requirements.txt
mcp
gradio
gradio[mcp]
gradio-client
openai
pandas
numpy
python-dotenv
cryptography